Add --list arg to list available scenarios + opts
This requires quite a lot a change to make NEOS works (in some well
defined cases) outside of job environment. The scenarios must also be
loaded with their opts sooner than before, eventually outside of job
context as well. This must be handled carefully with eventual job
related placeholders in defaults opts values.

The neos script can now launches the AppInEnv directly (skipping the
App class) if --module arg is not set. The logic is handled be the
Launcher which defines the best app to launch automatically.