Now handle logfile at Scenario level with --log
The feature to log commands outputs to a logfile is now managed at the
base Scenario class level. The logging is controlled with the -l, --log
flag on neos command line (disabled by default).

When the logging feature is not enabled, the commands outputs are
controlled by slurmstepd/srun which by default:
- redirect on stdout/stderr of the user tty on frontend nodes with
  salloc
- write in the jobs logfile(s) with sbatch

The behaviour of srun can also be controlled with --output and --error
options.