exec inenv in bash and forward SIGINT properly
When --module parameter is used and AppInEnv is launched in sub-process
by App, the sub-process is now executed with bash built-in exec to make
the shell exec the program without forking. This way, the bash process
disapears from the process tree and the main python program can easily
send signals to the AppInEnv sub-process. This mechanism is then used
to forward SIGINT to AppInEnv, which make it leave cleanly by killing
all its own sub-processes and removing all tmp files if needed.