Fix error handling if unknown opt in params
This commit delays the opts in parameters evaluation from scenario load
time to just before scenario run time. All scenarios are loaded on NEOS
launch but all do not accepts all possible opts. Only the run scenario
is supposed to accept the opt in param. This way, we avoid clueless
errors about unsupported opt in param when scenarios are loaded.

The exception name was also wrong, renamed from IndexError to KeyError
for proper catch.

The opts parsing errors are now properly handled: the appropriate error
message is printed and the opt is skipped.

Fixes #8