Allow nested srun to overlap parent srun
Recent versions of Slurm >= 20.11 do not allow by default srun in srun
to overlap on the same resources. The nested srun waits (endlessly
here) the parent srun to finish and release the resources. With neos, it
actually prevents the window manager to start in the neos job.

The previous srun behaviour can be obtained with --overlap parameter, to
allow the nested srun to use the same resources as the parent neos srun.