Release 0.6.17
With versions of dbus-launch greater or equal to 1.11.4, we can use
--exit-with-x11 as --exit-with-session is not recommended anymore (and
it our case, it can trigger a bug as the process won't die if it
couldn't connect to the X server).

We still need to support older versions of dbus-launch, hence we test
the output of 'dbus-launch --help' (which is easier than parsing the
version number).